Rename and move ErrorFactory to ErrorMapper.
